+++ src/Legend2Tool.WPF/MainWindow.xaml
@@ -1,21 +1,22 @@
-﻿<Window x:Class="Legend2Tool.WPF.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:hc="https://handyorg.github.io/handycontrol"
-        xmlns:local="clr-namespace:Legend2Tool.WPF"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:views="clr-namespace:Legend2Tool.WPF.Views"
-        xmlns:vm="clr-namespace:Legend2Tool.WPF.ViewModels"
-        Title="传奇配置工具"
-        Width="1070"
-        Height="700"
-        MinWidth="1070"
-        MinHeight="700"
-        d:DataContext="{d:DesignInstance vm:MainViewModel}"
-        hc:Growl.GrowlParent="True"
-        WindowStartupLocation="CenterOwner"
-        mc:Ignorable="d">
+﻿<Window
+    x:Class="Legend2Tool.WPF.MainWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:hc="https://handyorg.github.io/handycontrol"
+    xmlns:local="clr-namespace:Legend2Tool.WPF"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:views="clr-namespace:Legend2Tool.WPF.Views"
+    xmlns:vm="clr-namespace:Legend2Tool.WPF.ViewModels"
+    Title="传奇配置工具"
+    Width="1070"
+    Height="600"
+    MinWidth="1070"
+    MinHeight="600"
+    d:DataContext="{d:DesignInstance vm:MainViewModel}"
+    hc:Growl.GrowlParent="True"
+    WindowStartupLocation="CenterOwner"
+    mc:Ignorable="d">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
@@ -23,17 +24,19 @@
             <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
 
-        <views:MenuView Grid.Row="0"
-                        DataContext="{Binding MenuViewModel}"
-                        Grid.ColumnSpan="2" />
+        <views:MenuView
+            Grid.Row="0"
+            Grid.ColumnSpan="2"
+            DataContext="{Binding MenuViewModel}" />
 
-        <TabControl Grid.Row="1"
-                    HorizontalContentAlignment="Left"
-                    ItemsSource="{Binding ViewModels}"
-                    SelectedItem="{Binding SelectedViewModel}"
-                    Grid.ColumnSpan="2"
-                    Margin="0,0,0,30"
-                    Grid.RowSpan="2">
+        <TabControl
+            Grid.Row="1"
+            Grid.RowSpan="2"
+            Grid.ColumnSpan="2"
+            Margin="0,0,0,30"
+            HorizontalContentAlignment="Left"
+            ItemsSource="{Binding ViewModels}"
+            SelectedItem="{Binding SelectedViewModel}">
             <TabControl.ItemTemplate>
                 <DataTemplate>
                     <TextBlock Text="{Binding Head}" />
@@ -43,27 +46,31 @@
                 <DataTemplate DataType="{x:Type vm:PortConfViewModel}">
                     <views:PortConfView />
                 </DataTemplate>
+                <DataTemplate DataType="{x:Type vm:DynamicMonsterSpawningViewModel}">
+                    <views:DynamicMonsterSpawningView />
+                </DataTemplate>
                 <DataTemplate DataType="{x:Type vm:ScriptOptimizationViewModel}">
                     <views:ScriptOptimizationView />
                 </DataTemplate>
             </TabControl.Resources>
         </TabControl>
 
-        <Grid Grid.Row="2"
-              VerticalAlignment="Center"
-              Grid.ColumnSpan="2">
+        <Grid
+            Grid.Row="2"
+            Grid.ColumnSpan="2"
+            VerticalAlignment="Center">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="*" />
                 <ColumnDefinition Width="auto" />
             </Grid.ColumnDefinitions>
-            <ProgressBar Grid.Column="0"
-                         Value="{Binding ProgressPercentage}" />
-            <TextBlock Grid.Column="1"
-                       Margin="10,0,10,0"
-                       VerticalAlignment="Center"
-                       Foreground="Blue"
-                       MinWidth="200"
-                       Text="{Binding ProgressText}" />
+            <ProgressBar Grid.Column="0" Value="{Binding ProgressPercentage}" />
+            <TextBlock
+                Grid.Column="1"
+                MinWidth="200"
+                Margin="10,0,10,0"
+                VerticalAlignment="Center"
+                Foreground="Blue"
+                Text="{Binding ProgressText}" />
         </Grid>
     </Grid>
