+++ src/Legend2Tool.WPF/ViewModels/MainViewModel.cs
@@ -25,10 +25,10 @@ namespace Legend2Tool.WPF.ViewModels
 
         [ObservableProperty]
         private ViewModelBase? _selectedViewModel;
-        public MainViewModel(IDialogService dialogService, MenuViewModel menuViewModel, PortConfViewModel portConfViewModel, ScriptOptimizationViewModel scriptOptimizationViewModel, ProgressStore progressStore)
+        public MainViewModel(IDialogService dialogService, MenuViewModel menuViewModel, PortConfViewModel portConfViewModel, DynamicMonsterSpawningViewModel dynamicMonsterSpawningViewModel, ScriptOptimizationViewModel scriptOptimizationViewModel, ProgressStore progressStore)
         {
             MenuViewModel = menuViewModel;
-            ViewModels = new ObservableCollection<ViewModelBase> { portConfViewModel, scriptOptimizationViewModel };
+            ViewModels = new ObservableCollection<ViewModelBase> { portConfViewModel, dynamicMonsterSpawningViewModel, scriptOptimizationViewModel };
             SelectedViewModel = portConfViewModel;
             _progressStore = progressStore;
             WeakReferenceMessenger.Default.Register<ProgressChangedMessage>(this);
